PORT-4354
Make sure the dates for the attachments under Avail bundle are sorted (descendingly) by the "UpdatedAt" field if available or "received" date otherwise.

TODO: I am not able to find data samples in dev or QA (as in PORT-4278 5-13-2021 10:23 am that doesn't seem to exist anymore on QA).

TODO: lets devise a junit test instead.